Character no longer attached to parent of chain
Would probably cause issues in the future, if there's a mesh attached to
the character. In this case the mesh would show up twice (if both
character and actor are attached to the parent).
As of now, everything should always be attached to the Actor, and the
Actor is attached to the scene via the parent.